serpapi-keyword-extractor/requirements.txt
requests==2.31.0
nltk==3.8.1
python-dotenv==1.0.0